Move convertor to fhir package
